+++ src/test/java/dk/kb/datahandler/storage/JobStorageTest.java
@@ -31,7 +31,7 @@ public class JobStorageTest {
 
     @BeforeAll
     public static void beforeClass() throws Exception {
-        ServiceConfig.initialize("ds-datahandler-unittest.yaml");
+        ServiceConfig.initialize("conf/ds-datahandler-behaviour.yaml");
 
         H2DbUtil.createEmptyH2DBFromDDL(DB_URL, DRIVER, USERNAME, PASSWORD);
         JobStorage.initialize(DRIVER, DB_URL, USERNAME, PASSWORD);